add various updates
Removed timeout for set configuration **and** creation of a set.
Since rules to handle the set are required in any case, it makes
little sense to add one if not already present.

Added some defaults for syslog, if none is specified.

Added a default configuration file in '/etc/' if none is specified.