switch to intermittend file reading
This branch will explore the possibility to switch from keeping
an open file through the watch life to opening the file only
for reading.

By looking at the stats I noticed that, in my situation, the
number of events are quite limited.
Also, through _lsof_ I noticed that the binary had dangling
open files.
In this branch I will try to explore a solution to the above
through opening and closing the log files with each read.